+++ R/exp_shiny.R
@@ -1,6 +1,6 @@
 #' Interactively explore experience data
 #'
-#' @description Launch a shiny application to interactively explore drivers of
+#' @description Launch a Shiny application to interactively explore drivers of
 #' experience.
 #'
 #' `dat` must be an `exposed_df` object. An error will be thrown is any other
@@ -67,7 +67,10 @@
 #' - y: y variable
 #' - Geometry: plotting geometry
 #' - Add Smoothing?: activate to plot loess curves
-#' - Free y Scales: activate to enable separate y scales in each plot.
+#' - Second y-axis?: activate to enable a second y-axis
+#' - Second axis y: y variable to plot on the second axis
+#' - Free y Scales: activate to enable separate y scales in each plot
+#' - Log y-axis: activate to plot all y-axes on a log-10 scale
 #'
 #' ### Table
 #'
@@ -86,15 +89,18 @@
 #'
 #' @param dat An `exposed_df` object.
 #' @param predictors A character vector of independent variables in `dat` to
-#' include in the shiny app.
+#' include in the Shiny app.
 #' @param expected A character vector of expected values in `dat` to include
-#' in the shiny app.
+#' in the Shiny app.
 #' @param distinct_max Maximum number of distinct values allowed for `predictors`
 #' to be included as "Color" and "Facets" grouping variables. This input
 #' prevents the drawing of overly complex plots. Default value = 25.
+#' @param title Optional. Title of the Shiny app. If no title is provided,
+#' a descriptive title will be generated based on attributes of `dat`.
+#' @inheritParams exp_stats
 #'
 #' @return No return value. This function is called for the side effect of
-#' launching a shiny application.
+#' launching a Shiny application.
 #'
 #' @examples
 #'
@@ -103,10 +109,10 @@
 #'   expected_table <- c(seq(0.005, 0.03, length.out = 10), 0.2, 0.15, rep(0.05, 3))
 #'
 #'   study_py <- study_py |>
-#'   mutate(expected_1 = expected_table[pol_yr],
-#'                 expected_2 = ifelse(inc_guar, 0.015, 0.03)) |>
-#'   add_transactions(withdrawals) |>
-#'   left_join(account_vals, by = c("pol_num", "pol_date_yr"))
+#'     mutate(expected_1 = expected_table[pol_yr],
+#'            expected_2 = ifelse(inc_guar, 0.015, 0.03)) |>
+#'     add_transactions(withdrawals) |>
+#'     left_join(account_vals, by = c("pol_num", "pol_date_yr"))
 #'
 #'   exp_shiny(study_py)
 #' }
@@ -115,7 +121,11 @@
 exp_shiny <- function(dat,
                       predictors = names(dat),
                       expected = names(dat)[grepl("expected", names(dat))],
-                      distinct_max = 25L) {
+                      distinct_max = 25L,
+                      title,
+                      credibility = TRUE,
+                      cred_p = 0.95,
+                      cred_r = 0.05) {
 
   rlang::check_installed("shiny")
   rlang::check_installed("bslib")
@@ -142,28 +152,38 @@ exp_shiny <- function(dat,
   preds <- data.frame(predictors = predictors) |>
     # drop non-predictors (if any)
     filter(!predictors %in% c("pol_num", "status",
-                                     "term_date","exposure", trx_cols)) |>
+                              "term_date","exposure", trx_cols)) |>
     mutate(class1 = purrr::map_chr(predictors, ~ class(dat[[.x]])[[1]]),
-                  order = dplyr::case_when(
-                    class1 == "Date" ~ 1,
-                    class1 == "logical" ~ 2,
-                    class1 %in% c("character", "factor") ~ 3,
-                    class1 %in% c("numeric", "integer", "double") ~ 4,
-                    TRUE ~ 5
-                  ),
-                  is_number = purrr::map_lgl(predictors,
-                                             ~ is.numeric(dat[[.x]])),
-                  n_unique = purrr::map_int(predictors,
-                                            ~ dplyr::n_distinct(dat[[.x]]))) |>
+           order = dplyr::case_when(
+             class1 == "Date" ~ 1,
+             class1 == "logical" ~ 2,
+             class1 %in% c("character", "factor") ~ 3,
+             class1 %in% c("numeric", "integer", "double") ~ 4,
+             TRUE ~ 5
+           ),
+           is_number = purrr::map_lgl(predictors,
+                                      ~ is.numeric(dat[[.x]])),
+           n_unique = purrr::map_int(predictors,
+                                     ~ dplyr::n_distinct(dat[[.x]])),
+           scope = purrr::map2(predictors, class1,
+                               ~ if (.y %in% c("Date", "numeric",
+                                               "integer", "double")) {
+                                 range(dat[[.x]], na.rm = TRUE)
+                               } else {
+                                 unique(dat[[.x]])
+                               }
+           )
+    ) |>
     arrange(order) |>
     select(-order)
 
   preds_small <- filter(preds, n_unique <= distinct_max)$predictors
 
-  yVar_exp <- c("q_obs", "n_claims", "claims", "exposure", "credibility")
+  yVar_exp <- c("q_obs", "n_claims", "claims", "exposure",
+                if (credibility) "credibility")
   if (has_trx) {
     yVar_trx <- c("trx_util", "trx_freq", "trx_n", "trx_flag",
-                  "trx_amt", "avg_trx", "avg_all")
+                  "trx_amt", "avg_trx", "avg_all", "exposure")
     available_studies <- c("Termination study" = "exp",
                            "Transaction study" = "trx")
   } else {
@@ -176,6 +196,8 @@ exp_shiny <- function(dat,
                      checkbox_limit = 8) {
 
     inputId <- paste("i", x, sep = "_")
+    info <- filter(preds, predictors == x)
+    choices <- info$scope[[1]]
 
     if (is.null(dat[[x]])) {
       rlang::abort(
@@ -187,28 +209,24 @@ exp_shiny <- function(dat,
 
       shiny::sliderInput(
         inputId, shiny::strong(x),
-        min = min(dat[[x]], na.rm = TRUE),
-        max = max(dat[[x]], na.rm = TRUE),
-        value = range(dat[[x]], na.rm = TRUE)
+        min = choices[[1]],
+        max = choices[[2]],
+        value = choices
       )
 
     } else if (lubridate::is.Date(dat[[x]])) {
 
-      date_range <- range(dat[[x]], na.rm = TRUE)
-
       shiny::dateRangeInput(
         inputId, shiny::strong(x),
-        start = date_range[[1]],
-        end = date_range[[2]],
-        min = date_range[[1]],
-        max = date_range[[2]],
+        start = choices[[1]],
+        end = choices[[2]],
+        min = choices[[1]],
+        max = choices[[2]],
         startview = "year"
       )
 
     } else if (is.character(dat[[x]]) || is.logical(dat[[x]]) ||
                is.factor(dat[[x]])) {
-
-      choices <- unique(dat[[x]])
 
       if (length(choices) > checkbox_limit) {
         shiny::selectInput(
@@ -219,7 +237,7 @@ exp_shiny <- function(dat,
       } else {
         shiny::checkboxGroupInput(
           inputId, shiny::strong(x),
-          choices = unique(dat[[x]]), selected = choices
+          choices = choices, selected = choices
         )
       }
 
@@ -306,14 +324,18 @@ exp_shiny <- function(dat,
     trx_tab <- NULL
   }
 
+  if (missing(title)) {
+    title <- paste(attr(dat, "target_status"), collapse = "/") |>
+      paste("Experience Study",
+            if (has_trx) {
+              glue::glue("and {paste(all_trx_types, collapse = '/')} Transaction Study")})
+  }
+
   ui <- shiny::fluidPage(
 
     theme = bslib::bs_theme(bootswatch = "flatly"),
 
-    shiny::titlePanel(paste(attr(dat, "target_status"), collapse = "/") |>
-                        paste("Experience Study",
-                              if(has_trx) {
-                                glue::glue("and {paste(all_trx_types, collapse = '/')} Transaction Study")})),
+    shiny::titlePanel(title),
 
     shiny::sidebarLayout(
       shiny::sidebarPanel(
@@ -386,8 +408,19 @@ exp_shiny <- function(dat,
             shiny::fluidRow(
               shiny::column(
                 width = 4,
+                shiny::checkboxInput("plot2ndY",
+                                     shiny::strong("Second y-axis?"),
+                                     value = FALSE)
+              ),
+              selectPred("yVar_2nd", "Second axis y:", 4, choices = yVar_exp,
+                         selected = "exposure"),
+              shiny::column(
+                width = 4,
                 shiny::checkboxInput("plotFreeY",
                                      shiny::strong("Free y Scales?"),
+                                     value = FALSE),
+                shiny::checkboxInput("plotLogY",
+                                     shiny::strong("Log y-axis?"),
                                      value = FALSE)
               )
 
@@ -406,7 +439,7 @@ exp_shiny <- function(dat,
           )
         ),
 
-        shiny::h3("Filter Information"),
+        shiny::h3("Filter information"),
         shiny::verbatimTextOutput("filterInfo")
 
       )
@@ -421,7 +454,9 @@ exp_shiny <- function(dat,
       c(yVar_exp,
         input$ex_checks,
         glue::glue("ae_{input$ex_checks}"),
-        glue::glue("adj_{input$ex_checks}"))
+        glue::glue("adj_{input$ex_checks}"),
+        "All termination rates",
+        if (length(input$ex_checks) > 0) "All A/E ratios")
     })
 
     yVar_trx2 <- shiny::reactive({
@@ -433,22 +468,82 @@ exp_shiny <- function(dat,
     })
 
     # update y variable selections in response to inputs
-    shiny::observe(
+    shiny::observe({
+
+      new_choices <- if (input$study_type == "exp") {
+        yVar_exp2()
+      } else {
+        yVar_trx2()
+      }
+
       shiny::updateSelectInput(
-        session, "yVar", choices =
-          if(input$study_type == "exp") {
-            yVar_exp2()
-          } else {
-            yVar_trx2()
-          }
+        session, "yVar", choices = new_choices,
+        selected = if (input$yVar %in% new_choices) {
+          input$yVar
+        } else {
+          new_choices[[1]]
+        }
       )
-    ) |>
+
+    }) |>
       shiny::bindEvent(input$study_type, input$ex_checks, input$pct_checks)
+
+    shiny::observe({
+
+      if (input$study_type == "exp") {
+
+        new_choices_2 <- yVar_exp2()[
+          !yVar_exp2() %in% c("All termination rates", "All A/E ratios")]
+        if(input$yVar == "All termination rates") {
+          new_choices_2 <- new_choices_2[
+            !new_choices_2 %in% c('q_obs', input$ex_checks)]
+        } else if (input$yVar == "All A/E ratios") {
+          new_choices_2 <- new_choices_2[
+            !new_choices_2 %in% paste0("ae_", input$ex_checks)]
+        }
+
+      } else {
+        new_choices_2 <- yVar_trx2()
+      }
+
+      shiny::updateSelectInput(
+        session, "yVar_2nd", choices = new_choices_2,
+        selected = if (input$yVar_2nd %in% new_choices_2) {
+          input$yVar_2nd
+        } else {
+          "exposure"
+        }
+      )
+
+    }) |>
+      shiny::bindEvent(input$study_type, input$ex_checks, input$pct_checks,
+                       input$yVar)
+
+    # disable color input when using special plots
+    shiny::observe(
+      if (input$yVar %in% c("All termination rates", "All A/E ratios")) {
+        shiny::updateSelectInput(
+          session, "colorVar", choices = "Series", selected = "Series")
+      } else {
+        shiny::updateSelectInput(
+          session, "colorVar", choices = c("None", preds_small),
+          selected = if (input$colorVar %in% c("None", preds_small)) {
+            input$colorVar
+          } else {
+            "None"
+          }
+        )
+      }
+    ) |>
+      shiny::bindEvent(input$yVar, input$ex_checks)
 
     # reactive data
     rdat <- shiny::reactive({
 
-      filters <- purrr::map(preds$predictors, expr_filter)
+      keep <- purrr::imap_lgl(preds$predictors,
+                              ~ length(setdiff(preds$scope[[.y]],
+                                               input[[paste0("i_", .x)]])) > 0)
+      filters <- purrr::map(preds$predictors[keep], expr_filter)
 
       dat |>
         filter(!!!filters)
@@ -458,7 +553,7 @@ exp_shiny <- function(dat,
     rxp <- shiny::reactive({
 
       .groups <- c(input$xVar, input$colorVar, input$facetVar)
-      .groups <- .groups[.groups != "None"]
+      .groups <- .groups[!.groups %in% c("None", "Series")]
 
       if (input$weightVar == "None") {
         wt <- NULL
@@ -476,7 +571,8 @@ exp_shiny <- function(dat,
       if (input$study_type == "exp") {
         rdat() |>
           group_by(dplyr::across(dplyr::all_of(.groups))) |>
-          exp_stats(wt = wt, credibility = TRUE, expected = ex)
+          exp_stats(wt = wt, credibility = credibility, expected = ex,
+                    cred_p = cred_p, cred_r = cred_r)
       } else {
         rdat() |>
           group_by(dplyr::across(dplyr::all_of(.groups))) |>
@@ -489,8 +585,10 @@ exp_shiny <- function(dat,
 
     output$xpPlot <- shiny::renderPlot({
 
-      if (input$study_type == "exp" && input$yVar %in% yVar_trx2()) return()
-      if (input$study_type == "trx" && input$yVar %in% yVar_exp2()) return()
+      if (input$study_type == "exp" && input$yVar %in% yVar_trx2() &
+          !input$yVar == "exposure") return()
+      if (input$study_type == "trx" && input$yVar %in% yVar_exp2() &
+          !input$yVar == "exposure") return()
 
       dat <- rxp()
 
@@ -505,27 +603,49 @@ exp_shiny <- function(dat,
         rlang::sym(input$colorVar)
       }
 
-      y <- rlang::sym(input$yVar)
+      # set up y-variable and plotting function
+      if (input$yVar == "All termination rates") {
+        y <- rlang::expr(Rate)
+        plot.fun <- plot_termination_rates
+      } else if (input$yVar == "All A/E ratios") {
+        y <- rlang::expr(`A/E ratio`)
+        plot.fun <- plot_actual_to_expected
+      } else {
+        if (!all(c(input$yVar, as.character(color)) %in% names(rxp()))) return()
+        y <- rlang::sym(input$yVar)
+        plot.fun <- autoplot
+      }
+
+      second_y <- rlang::sym(input$yVar_2nd)
 
       mapping <- ggplot2::aes(!!x, !!y, color = !!color,
                               fill = !!color, group = !!color)
 
       # y labels
-      if (input$yVar %in% c("claims", "n_claims", "exposure",
-                            "trx_n", "trx_flag", "trx_amt",
-                            "avg_trx", "avg_all",
-                            input$pct_checks,
-                            paste0(input$pct_checks, "_w_trx"))) {
-        y_labels <- scales::label_comma(accuracy = 1)
-      } else if (input$yVar == "trx_freq") {
-        y_labels <- scales::label_comma(accuracy = 0.1)
-      } else {
-        y_labels <- scales::label_percent(accuracy = 0.1)
+      get_y_labels <- function(x) {
+        if (x %in% c("claims", "n_claims", "exposure",
+                     "trx_n", "trx_flag", "trx_amt",
+                     "avg_trx", "avg_all",
+                     input$pct_checks,
+                     paste0(input$pct_checks, "_w_trx"))) {
+          scales::label_comma(accuracy = 1)
+        } else if (x == "trx_freq") {
+          scales::label_comma(accuracy = 0.1)
+        } else {
+          scales::label_percent(accuracy = 0.1)
+        }
       }
 
+      y_labels <- get_y_labels(input$yVar)
+      second_y_labels <- get_y_labels(input$yVar_2nd)
+
       if (is.null(input$facetVar)) {
-        p <- dat |> autoplot(mapping = mapping, geoms = input$plotGeom,
-                             y_labels = y_labels)
+        p <- dat |> plot.fun(mapping = mapping, geoms = input$plotGeom,
+                             y_labels = y_labels,
+                             second_axis = input$plot2ndY,
+                             second_y = !!second_y,
+                             second_y_labels = second_y_labels,
+                             y_log10 = input$plotLogY)
       } else {
 
         facets <- rlang::syms(input$facetVar)
@@ -533,11 +653,15 @@ exp_shiny <- function(dat,
           facets <- append(facets, rlang::sym("trx_type"))
         }
 
-        p <- dat |> autoplot(!!!facets, mapping = mapping,
+        p <- dat |> plot.fun(!!!facets, mapping = mapping,
                              geoms = input$plotGeom,
                              y_labels = y_labels,
                              scales =
-                               if (input$plotFreeY) "free_y" else "fixed")
+                               if (input$plotFreeY) "free_y" else "fixed",
+                             second_axis = input$plot2ndY,
+                             second_y = !!second_y,
+                             second_y_labels = second_y_labels,
+                             y_log10 = input$plotLogY)
       }
 
       if (input$plotSmooth) p <- p + ggplot2::geom_smooth(method = "loess",
@@ -560,7 +684,7 @@ exp_shiny <- function(dat,
     output$filterInfo <- shiny::renderPrint({
       glue::glue("Total records = {scales::label_comma()(total_rows)}
                  Remaining records = {scales::label_comma()(nrow(rdat()))}
-                 % Data Remaining = {scales::label_percent(accuracy=0.1)(nrow(rdat())/total_rows)}")
+                 % Data remaining = {scales::label_percent(accuracy=0.1)(nrow(rdat())/total_rows)}")
     })
 
     output$xpDownload <- shiny::downloadHandler(