+++ R/exp_stats.R
@@ -30,31 +30,32 @@
 #' data while retaining any grouping variables passed to the "dots"
 #' (`...`).
 #'
-#' @param .data a data frame with exposure-level records, ideally of type `exposed_df`
-#' @param target_status a character vector of target status values
-#' @param expected a character vector containing column names in `.data`
+#' @param .data A data frame with exposure-level records, ideally of type `exposed_df`
+#' @param target_status A character vector of target status values
+#' @param expected A character vector containing column names in `.data`
 #' with expected values
-#' @param col_exposure name of the column in `.data` containing exposures
-#' @param col_status name of the column in `.data` containing the policy status
+#' @param col_exposure Name of the column in `.data` containing exposures
+#' @param col_status Name of the column in `.data` containing the policy status
 #' @param wt Optional. Length 1 character vector. Name of the column in
 #' `.data` containing weights to use in the calculation of claims,
 #' exposures, and partial credibility.
-#' @param credibility whether the output should include partial credibility
-#' weights and credibility-weighted decrement rates.
-#' @param cred_p confidence level under the Limited Fluctuation credibility method
-#' @param cred_r error tolerance under the Limited Fluctuation credibility
+#' @param credibility Whether the output should include partial credibility
+#' weights and credibility-weighted termination rates.
+#' @param cred_p Confidence level under the Limited Fluctuation credibility method
+#' @param cred_r Error tolerance under the Limited Fluctuation credibility
 #' method
-#' @param object an `exp_df` object
-#' @param ... groups to retain after `summary()` is called
+#' @param object An `exp_df` object
+#' @param ... Groups to retain after `summary()` is called
 #'
 #' @return A tibble with class `exp_df`, `tbl_df`, `tbl`,
 #' and `data.frame`. The results include columns for any grouping
-#' variables, claims, exposures, and observed decrement rates (`q_obs`).
+#' variables, claims, exposures, and observed termination rates (`q_obs`).
 #' If any values are passed to `expected`, additional columns will be
-#' added for expected decrements and actual-to-expected ratios. If
+#' added for expected termination rates and actual-to-expected ratios. If
 #' `credibility` is set to `TRUE`, additional columns are added
-#' for partial credibility and credibility-weighted decrement rates
-#' (assuming values are passed to `expected`).
+#' for partial credibility and credibility-weighted termination rates
+#' (assuming values are passed to `expected`). Credibility-weighted termination
+#' rates are prefixed by `adj_`.
 #'
 #' @examples
 #' toy_census |> expose("2020-12-31", target_status = "Surrender") |>
@@ -217,7 +218,7 @@ finish_exp_stats <- function(.data, target_status, expected,
         )))
     }
 
-    if(!is.null(expected)) {
+    if (!is.null(expected)) {
       adj_q_exp <- exp_form("credibility * q_obs + (1 - credibility) * {.col}",
                             "adj_{.col}", expected)
 
@@ -265,7 +266,16 @@ finish_exp_stats <- function(.data, target_status, expected,
 # Note - this could be handled using across, but is not due to performance on
 # grouped data frames
 exp_form <- function(form, new_col, .col) {
-  glue::glue(form) |>
+  gsub("\\{\\.col\\}", "`{.col}`", form) |>
+    glue::glue() |>
     purrr::set_names(glue::glue(new_col)) |>
     rlang::parse_exprs()
+}
+
+verify_exp_df <- function(.data) {
+  if (!inherits(.data, "exp_df")) {
+    rlang::abort(c(x = glue::glue("`{deparse(substitute(.data))}` must be an `exp_df` object."),
+                   i = "Hint: Use `exp_stats()` to create `exp_df` objects."
+    ))
+  }
 }