+++ R/globals.R
@@ -5,4 +5,5 @@ utils::globalVariables(c("issue_date", "term_date", "last_date",
                          "q_obs", ".weight", ".weight_n", ".weight_sq",
                          "ex_wt", "ex2_wt", "is_number", "n_unique",
                          "trx_date", "trx_type", "trx_amt", "trx_n", "trx_flag",
-                         "trx_util", "avg_trx", "avg_all", "trx_freq"))
+                         "trx_util", "avg_trx", "avg_all", "trx_freq",
+                         "class1", "Series", "Rate", "A/E ratio"))