+++ R/iam_basic_2012_g2.R
@@ -5,8 +5,8 @@
 #'
 #' @format For the 2012 IAMB table, a data frame with 242 rows and 3 columns:
 #' \describe{
-#'   \item{age}{attained age}
-#'   \item{qx}{mortality rate}
+#'   \item{age}{Attained age}
+#'   \item{qx}{Mortality rate}
 #'   \item{gender}{Female or Male}
 #' }
 #' @source
@@ -19,8 +19,8 @@
 #' @format
 #' For the Projection Scale G2 table, a data frame with 242 rows and 3 columns:
 #' \describe{
-#'   \item{age}{attained age}
-#'   \item{mi}{mortality improvement rate}
+#'   \item{age}{Attained age}
+#'   \item{mi}{Mortality improvement rate}
 #'   \item{gender}{Female or Male}
 #' }
 "scale_g2"