+++ tests/testthat/test-plots.R
@@ -55,19 +55,60 @@ test_that("Autoplot works with mapping overrides", {
 
   expect_s3_class(autoplot(exp_res4, inc_guar,
                            mapping = ggplot2::aes(x = pol_yr,
-                                                 y = ae_q_exp,
-                                                 fill = product),
+                                                  y = ae_q_exp,
+                                                  fill = product),
                            scales = "free_y",
                            geoms = "bars",
                            y_labels = scales::label_number()),
                   c("gg", "ggplot"))
   expect_s3_class(autoplot(trx_res4, trx_type, inc_guar,
                            mapping = ggplot2::aes(x = pol_yr,
-                                                 y = pct_of_premium_w_trx,
-                                                 fill = product),
+                                                  y = pct_of_premium_w_trx,
+                                                  fill = product),
                            scales = "free_y",
                            geoms = "bars",
                            y_labels = scales::label_number()),
                   c("gg", "ggplot"))
 
+})
+
+test_that("Second axis works", {
+  expect_s3_class(autoplot(exp_res4, second_axis = TRUE),
+                  c("gg", "ggplot"))
+  expect_s3_class(autoplot(trx_res4, second_axis = TRUE),
+                  c("gg", "ggplot"))
+  expect_s3_class(autoplot(trx_res4, second_axis = TRUE, second_y = trx_n),
+                  c("gg", "ggplot"))
+  expect_s3_class(autoplot(trx_res4, second_axis = TRUE, second_y = trx_n,
+                           second_y_labels = scales::comma(0.1)),
+                  c("gg", "ggplot"))
+})
+
+test_that("Termination plots works", {
+  expect_s3_class(plot_termination_rates(exp_res), c("gg", "ggplot"))
+  expect_s3_class(plot_termination_rates(exp_res, include_cred_adj = TRUE),
+                  c("gg", "ggplot"))
+  expect_error(plot_termination_rates(trx_res), regexp = "must be an `exp_df`")
+})
+
+test_that("AE plots works", {
+  expect_s3_class(plot_actual_to_expected(exp_res), c("gg", "ggplot"))
+  expect_error(plot_actual_to_expected(trx_res), regexp = "must be an `exp_df`")
+  expect_error(plot_actual_to_expected(expo |> exp_stats()),
+               regexp = "does not have any actual-to-expected")
+})
+
+test_that("Transaction utilization plots work", {
+  expect_s3_class(plot_utilization_rates(trx_res), c("gg", "ggplot"))
+  expect_s3_class(plot_utilization_rates(trx_res2), c("gg", "ggplot"))
+  expect_s3_class(plot_utilization_rates(trx_res3), c("gg", "ggplot"))
+  expect_s3_class(plot_utilization_rates(trx_res4), c("gg", "ggplot"))
+  expect_error(plot_utilization_rates(exp_res), regexp = "must be a `trx_df`")
+})
+
+test_that("Log y scale works", {
+  expect_s3_class(autoplot(exp_res4, y_log10 = TRUE, second_axis = TRUE),
+                  c("gg", "ggplot"))
+  expect_s3_class(autoplot(trx_res4, y_log10 = TRUE, second_axis = TRUE),
+                  c("gg", "ggplot"))
 })